CPU version of Resize operator (#127)
* CPU version of Resize operator
- new tests for CPU/GPU Resize operators added
- problem with the gray-scale image tests fixed


* New 24 tests for "Resize" operator added.
The changes requested by Janusz were made.


* Minor improvement in ResizeCropMirrorAttr class


* Changes suggested by Przemek


* Rollback of dali/CMakeLists.txt


* Pulled from latest upstream

+++ dali/pipeline/operators/decoder/host_decoder_test.cc
@@ -31,7 +31,7 @@ class HostDecodeTest : public GenericDecoderTest<ImgType> {
       .AddOutput("decoded", "cpu");
   }
 
-  uint8 GetTestCheckType() const  override {
+  uint32_t GetTestCheckType() const  override {
     return t_checkColorComp + t_checkElements;  // + t_checkAll + t_checkNoAssert;
   }
 };
